Move files and lock alert if no threshold has been set
