+++ frontend/src/components/MapView/LeftPanel/AlertsPanel/index.test.tsx
